split utils into setup and general
